refactor: complete low-priority Core improvements
- Consolidate UserDto: add UserRole Role + From(User) mapper to Core;
  delete WebApi/DTOs/UserDto.cs duplicate; update AdminController,
  ProtectedController, ProfileController, WebApp Admin view, AdminCli
  ConsoleUI and UsersMenu to use UserRole enum (no more int comparisons)
- Extract ApiClientBase to Core/Http/Impl/: implements all 6 IAppApiClient
  methods with shared helpers (SetBearer, ToJson, ReadAsync w/ class constraint);
  WebApp ApiClient inherits and adds ReadAsync logging override + 3 extra methods;
  AdminCli ApiClient inherits and overrides 6 base methods with try-catch + IAppLogger
- Tests: 303 passing (unchanged count; UserDtoTests updated for UserRole)

+++ src/AppSimple.WebApp/Controllers/AdminController.cs
@@ -86,7 +86,7 @@ public sealed class AdminController : Controller
             PhoneNumber = user.PhoneNumber,
             Bio = user.Bio,
             DateOfBirth = user.DateOfBirth,
-            Role = user.Role,
+            Role = (int)user.Role,
             IsActive = user.IsActive
         });
     }
